requirements.txt
ollama
streamlit
langchain_community
langchain_core
faiss-cpu
sentence-transformers
unstructured
unstructured[pdf]
unstructured[docx]
opencv-python-headless
langchain-chroma
chromadb==0.3.29